documentation and some minor fixes
