+++ src/index.ts
@@ -5,3 +5,4 @@ export * from "./utility";
 
 export * from "./schema";
 export * from "./validation";
+export * from "./refine";
